REFACTOR: calculateMetrics.js renamed for consistency functionality contained within  App.jsx & calculateMetrics.js
